% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UVVisDataZnO}
\alias{UVVisDataZnO}
\title{UV-Vis spectra of growing, colloidal ZnO quantum dots}
\format{
A data frame with 67257 rows and 12 variables:
\describe{
  \item{sampleid}{A string identifying the sample/spectrum}
  \item{time}{Time since start of data collection, in minutes}
  \item{wavelength}{Wavelength/nm}
  \item{intensity}{Absorbance, unitless}
  \item{IntegrationTime}{}
  \item{n_Averaged}{}
  \item{Boxcar}{}
  \item{CorrElectricDark}{}
  \item{StrobeLampEnabled}{}
  \item{CorrDetectorNonLin}{}
  \item{CorrStrayLight}{}
  \item{n_Pixels}{}
}
}
\source{
\url{https://doi.org/10.1021/acs.jpcc.9b11229}

\url{https://solarchemist.se/2021/08/02/composite-gif-gganimate/}
}
\usage{
UVVisDataZnO
}
\description{
Experimental dataset of 141 UV-Vis spectra.
    Recorded using OceanOptics HR2000 UV-Vis spectrophotometer.
    For more details, see the cited sources.
}
\keyword{datasets}
