% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauc.R
\name{tauc}
\alias{tauc}
\title{Calculates optical band gap using Tauc plot}
\usage{
tauc(
  energy,
  absorbance,
  r = 0.5,
  lowE.limits,
  highE.limits,
  bg.limits = c(0.3, 0.8)
)
}
\arguments{
\item{energy, }{numeric vector, in eV}

\item{absorbance, }{numeric vector, unitless, should be same length as energy}

\item{r, }{exponent that determines transition type, 0.5 for direct allowed (default),
1.5 for direct forbidden, 2 for indirect allowed, and 3 for indirect forbidden transitions.}

\item{lowE.limits, }{vector of two numbers, defines low-energy plateau
two energy values, that define a straight line approximating the low-energy plateau}

\item{highE.limits, }{vector of two numbers, defines high-energy plateau
two energy values, that define a straight line approximating the high-energy plateau}

\item{bg.limits, }{vector of two numbers, relative vertical bounds defining the lower and upper extent
of points in the band edge to fit (default is c(0.3, 0.8))}
}
\value{
tibble with same number of rows as length of input vectors,
    with the following columns:
    + energy (input values returned)
    + absorbance (input values returned)
    + ceiling, line that defines the "ceiling" that this function used
    + floor, line that defines the "ceiling" that this function used
    + edge, boolean, TRUE for the rows that are part of the Tauc fit
    + fit.Eg, the calculated optical band gap
    + fit.tauc, the fitted Tauc line
    + fit.adj_rsq, adjusted R-square of the linear Tauc fit
    + fit.points, the number of data points that are part of the Tauc fit (cf. edge)
}
\description{
This function uses Tauc plotting to determine the optical band gap
from your supplied UV-Vis absorbance spectra.
}
\details{
For the next version of this function, consider how we could
adapt it to handle energy/absorbance with associated errors (error propagation).
Actually, more importantly, make it handle multiple spectra first.
}
\examples{
\dontrun{
tauc(energy, absorbance, lowE.limits = c(2.4, 3.4), highE.limits = c(3.9, 4.3))
tauc(energy, absorbance, r = 2, lowE.limits = c(2.2, 3.0), highE.limits = c(3.4, 3.8))
tauc(energy, absorbance, bg.limits = c(0.20, 0.85), lowE.limits = ..., highE.limits = ...)
}
}
